+++ src/bika/lims/browser/analyses/view.py
@@ -871,9 +871,7 @@ class AnalysesView(ListingView):
 
         item["Result"] = ""
 
-        # TODO: The permission `ViewResults` is managed on the sample.
-        #       -> Change to a proper field permission!
-        if not self.has_permission(ViewResults, self.context):
+        if not self.has_permission(ViewResults, analysis_brain):
             # If user has no permissions, don"t display the result but an icon
             img = get_image("to_follow.png", width="16px", height="16px")
             item["before"]["Result"] = img