+++ src/senaite/core/z3cform/widgets/queryselect/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+from .widget import QuerySelectWidget  # noqa: F401
+from .widget import QuerySelectWidgetFactory  # noqa: F401