+++ CHANGES.rst
@@ -5,6 +5,11 @@ Changelog
 ------------------
 
 - #2120 Support for Multiple component analysis
+- #2180 Rely on analysis permission when displaying results
+- #2179 Fix Traceback when removing a Worksheet
+- #2178 AT Queryselect Widget
+- #2177 Dexterity Queryselect Widget
+- #2173 Fix UnicodeDecodeError when a required condition is empty in add form
 - #2172 Fix SuperModel import error introduced in PR #2154
 - #2169 Fix field error indication in sample header
 - #2166 Fix partitions not displaying complete list of Interpretation Templates